chap-2/requirements.txt
RPi.bme280
smbus2
